feat: types on fields view
