+++ src/components/Services/VariablesDisplay/VariablesList/ViewFields/ArrayList.tsx
@@ -2,14 +2,15 @@ import { useState } from 'react'
 import { classNames } from 'src/utils'
 import { useVariablesContext } from '../../VariablesContext'
 
-interface VariableFieldType {
+export function ArrayList({
+  prop,
+  value,
+  parent,
+}: {
   prop: string
-  value: string | number | boolean | any[]
+  value: any[]
   parent?: string
-}
-
-//TODO: Impose value as string
-export function StringNumberField({ prop, value, parent }: VariableFieldType) {
+}) {
   const { setNewVariables } = useVariablesContext()
   const [error, setError] = useState(false)
   const inputName = parent ? [parent, prop].join('.') : prop
@@ -33,22 +34,23 @@ export function StringNumberField({ prop, value, parent }: VariableFieldType) {
     }
   }
 
-  if (typeof value === 'string' || typeof value === 'number') {
-    return (
-      <div className="flex">
-        <input
-          name={inputName}
-          className={classNames(
-            'flex-grow bg-gray-100',
-            error && 'bg-red-200',
-            typeof value === 'number' ? 'text-teal-600' : 'text-slate-600'
-          )}
-          defaultValue={JSON.stringify(value)}
-          onChange={handleChange}
-        />
-      </div>
-    )
-  } else {
-    return <></>
-  }
+  return (
+    <ol className="flex flex-grow flex-col gap-2">
+      {value.map((v) => (
+        <li key={inputName} className="flex grow">
+          <input
+            name={inputName}
+            className={classNames(
+              'grow',
+              error && 'bg-red-200',
+              typeof value === 'number' ? 'text-teal-600' : 'text-slate-700',
+              'hover:opacity-100 hover:bg-slate-200 transition duration-75 ease-in-out'
+            )}
+            defaultValue={JSON.stringify(v)}
+            onChange={handleChange}
+          />
+        </li>
+      ))}
+    </ol>
+  )
 }
